+++ inc/Engine/Optimization/RUCSS/Warmup/ResourceFetcherProcess.php
@@ -4,11 +4,18 @@ declare(strict_types=1);
 namespace WP_Rocket\Engine\Optimization\RUCSS\Warmup;
 
 use WP_Rocket\Admin\Options;
+use WP_Rocket\Dependencies\Minify\CSS as MinifyCSS;
+use WP_Rocket\Dependencies\Minify\JS as MinifyJS;
+use WP_Rocket\Engine\Optimization\AssetsLocalCache;
+use WP_Rocket\Engine\Optimization\CSSTrait;
 use WP_Rocket\Engine\Optimization\RUCSS\Database\Queries\ResourcesQuery;
+use WP_Rocket\Engine\Optimization\UrlTrait;
 use WP_Rocket\Logger\Logger;
 use \WP_Rocket_WP_Background_Process;
 
 class ResourceFetcherProcess extends WP_Rocket_WP_Background_Process {
+
+	use UrlTrait, CSSTrait;
 
 	/**
 	 * Background process action name.
@@ -54,50 +61,71 @@ class ResourceFetcherProcess extends WP_Rocket_WP_Background_Process {
 	private $options_api;
 
 	/**
+	 * Assets local cache instance
+	 *
+	 * @var AssetsLocalCache
+	 */
+	private $local_cache;
+
+	/**
 	 * ResourceFetcherProcess constructor.
 	 *
-	 * @param ResourcesQuery $resources_query ResourcesQuery instance.
-	 * @param APIClient      $api_client APIClient instance.
-	 * @param Options        $options_api Options API instance.
+	 * @param ResourcesQuery   $resources_query ResourcesQuery instance.
+	 * @param APIClient        $api_client APIClient instance.
+	 * @param Options          $options_api Options API instance.
+	 * @param AssetsLocalCache $local_cache Local cache instance.
 	 */
-	public function __construct( ResourcesQuery $resources_query, APIClient $api_client, Options $options_api ) {
+	public function __construct( ResourcesQuery $resources_query, APIClient $api_client, Options $options_api, AssetsLocalCache $local_cache ) {
 		parent::__construct();
 
 		$this->resources_query = $resources_query;
 		$this->api_client      = $api_client;
 		$this->options_api     = $options_api;
+		$this->local_cache     = $local_cache;
 	}
 
 	/**
-	 * Get url file contents.
+	 * Minify and prepare CSS.
 	 *
-	 * @param string $url File url.
+	 * @param string $path Path of the CSS file.
+	 * @param string $contents Contents of the CSS file.
+	 *
+	 * @return string
+	 */
+	private function prepare_css_content( string $path, string $contents ) : string {
+		$contents = trim( $this->rewrite_paths( $path, $path, $contents ) );
+		$minifier = new MinifyCSS( $contents );
+
+		return $minifier->minify();
+	}
+
+	/**
+	 * Minify and prepare JS.
+	 *
+	 * @param string $contents Contents of the JS file.
+	 *
+	 * @return string
+	 */
+	private function prepare_js_content( string $contents ) : string {
+		$minifier = new MinifyJS( $contents );
+
+		return $minifier->minify();
+	}
+
+	/**
+	 * Get file contents by path.
+	 *
+	 * @param string $path File path.
 	 * @param string $type File type (css,js).
 	 *
-	 * @return array
+	 * @return string
 	 */
-	private function get_url_contents( $url, string $type = 'css' ) : array {
-		$external_url = $this->is_external_file( $url );
-
-		$file_path = $external_url ? $this->local_cache->get_filepath( $url ) : $this->get_file_path( $url );
-
-		if ( empty( $file_path ) ) {
-			Logger::error(
-				'Couldn’t get the file path from the URL.',
-				[
-					'RUCSS warmup process',
-					'url' => $url,
-				]
-			);
-
-			return [ md5( uniqid() ), '*' ];
-		}
-
-		$file_content = $external_url ? $this->local_cache->get_content( $url ) : $this->get_file_content( $file_path );
+	private function get_path_contents( string $path, string $type = 'css' ) : string {
+		$file_content = $this->get_file_content( $path );
 
 		// Minify the content if it's there.
 		if ( $file_content ) {
-			$file_content = 'js' === $type ? $this->prepare_js_content( $file_content ) : $this->prepare_css_content( $file_path, $file_content );
+			$file_content = 'js' === $type ? $this->prepare_js_content( $file_content ) : $this->prepare_css_content( $path, $file_content );
 		}
 
 		if ( ! $file_content ) {
@@ -105,14 +133,14 @@ class ResourceFetcherProcess extends WP_Rocket_WP_Background_Process {
 				'No file content.',
 				[
 					'RUCSS warmup process',
-					'path' => $file_path,
+					'path' => $path,
 				]
 			);
 
-			return [ md5( uniqid() ), '*' ];
+			return '*';
 		}
 
-		return [ $file_path, $file_content ];
+		return $file_content;
 	}
 
 	/**
@@ -152,6 +180,8 @@ class ResourceFetcherProcess extends WP_Rocket_WP_Background_Process {
 		if ( empty( $resource['url'] ) ) {
 			return false;
 		}
+
+		$resource['content'] = $this->get_path_contents( $resource['path'], $resource['type'] );
 
 		if ( $this->resources_query->create_or_update( $resource ) ) {
 			$this->content_changed = true;
